morpheme loader base version

+++ setup_project.py
@@ -6,26 +6,49 @@
 import urllib.request
 from itertools import product
 import subprocess
-
+import os
+from os import path
 from tqdm import tqdm
 
 if __name__ == '__main__':
     # Download Hebrew ELMo
-    print("Beginning Hebrew ELMo download...")
-    url = 'http://vectors.nlpl.eu/repository/11/154.zip'
-    urllib.request.urlretrieve(url, 'ELMoForManyLangs/hebrew.zip')
-    print("Finished downloading")
+    if path.isfile('ELMoForManyLangs/hebrew.zip'):
+        print('ELMoForManyLangs was already downloaded, moving on...')
+    else:
+        print("Beginning Hebrew ELMo download...")
+        url = 'http://vectors.nlpl.eu/repository/11/154.zip'
+        urllib.request.urlretrieve(url, 'ELMoForManyLangs/hebrew.zip')
+        print("Finished downloading")
 
     # Download Hebrew Tree Bank
-    print("Beginning Hebrew Tree Bank download...")
-    root_url = 'https://raw.github.com/OnlpLab/HebrewResources/master/HebrewTreebank/hebtb/'
-    for set, data in tqdm(
-            list(product(['train', 'dev', 'test'], ['-gold.conll', '-gold.lattices', '.lattices', '.tokens'])),
-            desc='Downloading..', unit='file'):
-        file = set + '_hebtb' + data
-        url = root_url + file
-        urllib.request.urlretrieve(url, f'data/hebrew_tree_bank/{file}')
-    print("Finished downloading")
+    if path.isdir('data/hebrew_tree_bank'):
+        print('HebrewTreebank already exists, moving on...')
+    else:
+        print("Beginning Hebrew Tree Bank download...")
+        root_url = 'https://raw.github.com/OnlpLab/HebrewResources/master/HebrewTreebank/hebtb/'
+        for set, data in tqdm(
+                list(product(['train', 'dev', 'test'], ['-gold.conll', '-gold.lattices', '.lattices', '.tokens'])),
+                desc='Downloading..', unit='file'):
+            file = set + '_hebtb' + data
+            url = root_url + file
+            urllib.request.urlretrieve(url, f'data/hebrew_tree_bank/{file}')
+        print("Finished downloading")
+
+    if path.isdir('yapproj'):
+        print('Yap already cloned, moving on...')
+    else:
+        os.mkdir('yapproj')
+        os.mkdir('yapproj/src')
+        print("Beginning to Download yap tool...")
+        subprocess.run(["git", "clone", "https://github.com/OnlpLab/yap.git","yapproj/src/yap"])
+        #TODO: extract data
+        #TODO: The following:
+        # 1. Set environment variable GOPATH=yapproj
+        # 2. go to the src directory
+        # 3. Do "go get ." and then "go build ."
+        # 0. Verify go installation
+
+        print("Finished cloning")
 
     # install the ELMO package
     subprocess.run("python ELMoForManyLangs/setup.py install")
